+++ src/imio/project/pst/migrations/migrate_to_1_3_3.py
@@ -4,6 +4,9 @@ import logging
 
 from collective.contact.plonegroup.config import FUNCTIONS_REGISTRY
 from imio.migrator.migrator import Migrator
+from imio.project.core.utils import getProjectSpace
+from imio.project.pst.setuphandlers import _ as _translate, COLUMNS_FOR_CONTENT_TYPES, createDashboardCollections
+from plone import api
 from plone.registry.interfaces import IRegistry
 from zope.component import getUtility
 
@@ -23,6 +26,36 @@ def migrate_plonegroups():
         registry[FUNCTIONS_REGISTRY] = registry[FUNCTIONS_REGISTRY] + to_add
 
 
+def add_representative_responsible_collection():
+    """Added an additional collection "Which I am representative responsible" for operational objective and pstaction"""
+    pst = api.portal.get().pst
+    for tup in [('operationalobjectives', ['operationalobjective']), ('pstactions', ['pstaction'])]:
+        col_folder = tup[0]
+        content_type = tup[1]
+        folder = pst[col_folder]
+        ps_path = '/'.join(getProjectSpace(folder).getPhysicalPath())
+        collections = [
+            {
+                'id': 'i-am-representative_responsible',
+                'tit': _translate("Which I am representative responsible"),
+                'subj': ('search',),
+                'query': [
+                    {'i': 'portal_type', 'o': 'plone.app.querystring.operation.selection.is', 'v': content_type},
+                    {'i': 'CompoundCriterion', 'o': 'plone.app.querystring.operation.compound.is',
+                     'v': 'user-is-representative-responsible'},
+                    {'i': 'path', 'o': 'plone.app.querystring.operation.string.path', 'v': ps_path}
+                ],
+                'cond': u"",
+                'bypass': [],
+                'flds': COLUMNS_FOR_CONTENT_TYPES[content_type[0]],
+                'sort': u'sortable_title',
+                'rev': False,
+                'count': False
+            },
+        ]
+        createDashboardCollections(folder, collections, 1)
+
+
 class MigrateTo133(Migrator):
 
     def __init__(self, context):
@@ -31,6 +64,9 @@ class MigrateTo133(Migrator):
     def run(self):
         # Added a new representative responsible function
         migrate_plonegroups()
+
+        # Added an additional collection "Which I am representative responsible" for operational objective and pstaction
+        add_representative_responsible_collection()
 
         # Display duration
         self.finish()